Added a cell_text field in semiauto, so that you can see what a cell's index is and also select a particular cell when needed.
The changes to CellGraph here are not useful-- I attempted to do some fancy stuff with finding which parts of the mesh are "empty" but it didn't really work out, so I commented it all out for now. Instead, if you try to remove an edge between a cell and a nonexistant cell, it just gives you a messagebox error.