autoAddEdge now tries to maximize a weighted sum of 	1) overlap 	2) the minimum angle created by the cell (so this is some sort of maximin except also with overlap)
I do this because sometimes it chooses badly by picking things that generate very small angles. at the moment the two are weighted with a hard-coded factor of 1, where the overlap is fractional (overlap score) and the min angle is in radians. thus the angle contributes fairly substantially. I should make this a GUI-tunable parameter later if i want to keep it.